requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
aiosqlite==0.19.0
requests==2.31.0
python-dotenv==1.0.0 